+++ tests/main.R
@@ -1,3 +1,5 @@
+# ERGYAPLÉCC/Erzsiplétsz
+
 # Needed
 ## install.packages(NULL
 ## 	, "ncdf4"