+++ tests/offline/test_offline_sf_multilines.R
@@ -23,7 +23,7 @@ expect_equal(st_crs(rec0)[1][[1]], proj)
 
 ## ################################################################################
 # Reconstructed to 100 
-rec100 <- reconstruct(admin, age=100, model=model, plateperiod=TRUE)
+rec100 <- reconstruct(admin, age=100, model=model, validtime=TRUE)
 
 
 
@@ -37,8 +37,8 @@ expect_equal(colnames(rec0), colnames(rec100))
 expect_equal(st_crs(rec100)[1][[1]], proj)
 
 ################################################################################
-# Reconstructed to 100 with plateperiod=FALSE
-rec100_pp <- reconstruct(admin, age=100, model=model, plateperiod=FALSE)
+# Reconstructed to 100 with validtime=FALSE
+rec100_pp <- reconstruct(admin, age=100, model=model, validtime=FALSE)
 
 
 expect_inherits(rec100_pp, "sf")
@@ -53,7 +53,7 @@ expect_equal(st_crs(rec100_pp)[1][[1]], proj)
 
 ################################################################################
 # Reconstructed to c(0,100) with listout
-rec <- reconstruct(admin, age=c(0,100), listout=TRUE,  model=model, plateperiod=TRUE)
+rec <- reconstruct(admin, age=c(0,100), listout=TRUE,  model=model, validtime=TRUE)
 
 
 # high-level attributes
@@ -67,7 +67,7 @@ expect_equal(rec[[2]], rec100)
 ################################################################################
 # Reconstructed to c(0,100) with listout=FALSE
 expect_error(
-	rec <- reconstruct(admin, age=c(0,100), listout=FALSE,  model=model, plateperiod=TRUE)
+	rec <- reconstruct(admin, age=c(0,100), listout=FALSE,  model=model, validtime=TRUE)
 )
 
 
@@ -98,7 +98,7 @@ expect_equal(st_crs(rec0)[1][[1]], proj)
 
 ## ################################################################################
 # Reconstructed to 100 
-rec100 <- reconstruct(admin, age=100, model=model, plateperiod=TRUE)
+rec100 <- reconstruct(admin, age=100, model=model, validtime=TRUE)
 
 
 
@@ -112,8 +112,8 @@ expect_equal(colnames(rec0), colnames(rec100))
 expect_equal(st_crs(rec100)[1][[1]], proj)
 
 ################################################################################
-# Reconstructed to 100 with plateperiod=FALSE
-rec100_pp <- reconstruct(admin, age=100, model=model, plateperiod=FALSE)
+# Reconstructed to 100 with validtime=FALSE
+rec100_pp <- reconstruct(admin, age=100, model=model, validtime=FALSE)
 
 
 expect_inherits(rec100_pp, "sf")
@@ -128,7 +128,7 @@ expect_equal(st_crs(rec100_pp)[1][[1]], proj)
 
 ################################################################################
 # Reconstructed to c(0,100) with listout
-rec <- reconstruct(admin, age=c(0,100), listout=TRUE,  model=model, plateperiod=TRUE)
+rec <- reconstruct(admin, age=c(0,100), listout=TRUE,  model=model, validtime=TRUE)
 
 
 # high-level attributes
@@ -142,6 +142,6 @@ expect_equal(rec[[2]], rec100)
 ################################################################################
 # Reconstructed to c(0,100) with listout=FALSE
 expect_error(
-	rec <- reconstruct(admin, age=c(0,100), listout=FALSE,  model=model, plateperiod=TRUE)
+	rec <- reconstruct(admin, age=c(0,100), listout=FALSE,  model=model, validtime=TRUE)
 )