+++ tests/offline/test_offline_sf_multipoint.R
@@ -46,11 +46,11 @@ expect_equal(st_crs(rec0)[1][[1]], "WGS 84")
 
 ## ################################################################################
 # Reconstructed to 100 
-rec100 <- reconstruct(sfp, age=100, model=model, plateperiod=TRUE)
+rec100 <- reconstruct(sfp, age=100, model=model, validtime=TRUE)
 
 
 # the data.frame method for cross checking
-recDF<- reconstruct(dmat[, c("lng", "lat")], age=100, model=model, plateperiod=TRUE)
+recDF<- reconstruct(dmat[, c("lng", "lat")], age=100, model=model, validtime=TRUE)
 
 
 expect_inherits(rec100, "sf")
@@ -74,8 +74,8 @@ expect_equal(part, full[names(part)])
 expect_equal(st_crs(rec100)[1][[1]], "WGS 84")
 
 ################################################################################
-# Reconstructed to 100 with plateperiod=FALSE
-rec100_pp <- reconstruct(sfp, age=100, model=model, plateperiod=FALSE)
+# Reconstructed to 100 with validtime=FALSE
+rec100_pp <- reconstruct(sfp, age=100, model=model, validtime=FALSE)
 
 
 expect_inherits(rec100_pp, "sf")
@@ -95,7 +95,7 @@ expect_equal(st_crs(rec100_pp)[1][[1]], "WGS 84")
 
 ################################################################################
 # Reconstructed to c(0,100) with listout
-rec <- reconstruct(sfp, age=c(0,100), listout=TRUE,  model=model, plateperiod=TRUE)
+rec <- reconstruct(sfp, age=c(0,100), listout=TRUE,  model=model, validtime=TRUE)
 
 
 # high-level attributes
@@ -110,7 +110,7 @@ expect_equal(rec[[2]], rec100)
 ################################################################################
 # Reconstructed to c(0,100) with listout=FALSE
 expect_error(
-	rec <- reconstruct(sfp, age=c(0,100), listout=FALSE,  model=model, plateperiod=TRUE)
+	rec <- reconstruct(sfp, age=c(0,100), listout=FALSE,  model=model, validtime=TRUE)
 )
 
 
@@ -146,11 +146,11 @@ expect_equal(st_crs(rec0)[1][[1]], proj)
 
 ## ################################################################################
 # Reconstructed to 100 
-rec100 <- reconstruct(sfp, age=100, model=model, plateperiod=TRUE)
+rec100 <- reconstruct(sfp, age=100, model=model, validtime=TRUE)
 
 
 # the data.frame method for cross checking
-recDF<- reconstruct(dmat[, c("lng", "lat")], age=100, model=model, plateperiod=TRUE)
+recDF<- reconstruct(dmat[, c("lng", "lat")], age=100, model=model, validtime=TRUE)
 
 
 expect_inherits(rec100, "sf")
@@ -174,8 +174,8 @@ expect_equal(part, full[names(part)])
 expect_equal(st_crs(rec100)[1][[1]], proj)
 
 ################################################################################
-# Reconstructed to 100 with plateperiod=FALSE
-rec100_pp <- reconstruct(sfp, age=100, model=model, plateperiod=FALSE)
+# Reconstructed to 100 with validtime=FALSE
+rec100_pp <- reconstruct(sfp, age=100, model=model, validtime=FALSE)
 
 
 expect_inherits(rec100_pp, "sf")
@@ -195,7 +195,7 @@ expect_equal(st_crs(rec100_pp)[1][[1]], proj)
 
 ################################################################################
 # Reconstructed to c(0,100) with listout
-rec <- reconstruct(sfp, age=c(0,100), listout=TRUE,  model=model, plateperiod=TRUE)
+rec <- reconstruct(sfp, age=c(0,100), listout=TRUE,  model=model, validtime=TRUE)
 
 
 # high-level attributes
@@ -209,7 +209,7 @@ expect_equal(rec[[2]], rec100)
 ################################################################################
 # Reconstructed to c(0,100) with listout=FALSE
 expect_error(
-	rec <- reconstruct(sfp, age=c(0,100), listout=FALSE,  model=model, plateperiod=TRUE)
+	rec <- reconstruct(sfp, age=c(0,100), listout=FALSE,  model=model, validtime=TRUE)
 )
 
 
@@ -246,11 +246,11 @@ expect_equal(st_crs(rec0)[1][[1]], proj)
 
 ## ################################################################################
 # Reconstructed to 100 
-rec100 <- reconstruct(sfp, age=100, model=model, plateperiod=TRUE)
+rec100 <- reconstruct(sfp, age=100, model=model, validtime=TRUE)
 
 
 # the data.frame method for cross checking
-recDF<- reconstruct(dmat[, c("lng", "lat")], age=100, model=model, plateperiod=TRUE)
+recDF<- reconstruct(dmat[, c("lng", "lat")], age=100, model=model, validtime=TRUE)
 
 
 expect_inherits(rec100, "sf")
@@ -274,8 +274,8 @@ expect_equal(part, full[names(part)])
 expect_equal(st_crs(rec100)[1][[1]], proj)
 
 ################################################################################
-# Reconstructed to 100 with plateperiod=FALSE
-rec100_pp <- reconstruct(sfp, age=100, model=model, plateperiod=FALSE)
+# Reconstructed to 100 with validtime=FALSE
+rec100_pp <- reconstruct(sfp, age=100, model=model, validtime=FALSE)
 
 
 expect_inherits(rec100_pp, "sf")
@@ -295,7 +295,7 @@ expect_equal(st_crs(rec100_pp)[1][[1]], proj)
 
 ################################################################################
 # Reconstructed to c(0,100) with listout
-rec <- reconstruct(sfp, age=c(0,100), listout=TRUE,  model=model, plateperiod=TRUE)
+rec <- reconstruct(sfp, age=c(0,100), listout=TRUE,  model=model, validtime=TRUE)
 
 
 # high-level attributes
@@ -309,7 +309,7 @@ expect_equal(rec[[2]], rec100)
 ################################################################################
 # Reconstructed to c(0,100) with listout=FALSE
 expect_error(
-	rec <- reconstruct(sfp, age=c(0,100), listout=FALSE,  model=model, plateperiod=TRUE)
+	rec <- reconstruct(sfp, age=c(0,100), listout=FALSE,  model=model, validtime=TRUE)
 )
 
 ########################################----------------------------------------
@@ -317,7 +317,7 @@ expect_error(
 
 ################################################################################
 # Reconstructed to c(0,100) with gmeta=TRUE
-rec100_pp_meta <- reconstruct(sfp, age=100, model=model, plateperiod=FALSE, gmeta=TRUE)
+rec100_pp_meta <- reconstruct(sfp, age=100, model=model, validtime=FALSE, gmeta=TRUE)
 
 
 expect_equal(rec100_pp, rec100_pp_meta[, -(1:4)])