+++ tests/offline/test_offline_sp_lines.R
@@ -24,7 +24,7 @@ expect_equal(st_crs(rec0)[1][[1]], proj)
 
 ## ################################################################################
 # Reconstructed to 100 
-rec100 <- reconstruct(admin, age=100, model=model, plateperiod=TRUE)
+rec100 <- reconstruct(admin, age=100, model=model, validtime=TRUE)
 
 
 
@@ -38,8 +38,8 @@ expect_equal(colnames(rec0), colnames(rec100))
 expect_equal(st_crs(rec100)[1][[1]], proj)
 
 ################################################################################
-# Reconstructed to 100 with plateperiod=FALSE
-rec100_pp <- reconstruct(admin, age=100, model=model, plateperiod=FALSE)
+# Reconstructed to 100 with validtime=FALSE
+rec100_pp <- reconstruct(admin, age=100, model=model, validtime=FALSE)
 
 
 expect_inherits(rec100_pp, "Spatial")
@@ -54,7 +54,7 @@ expect_equal(st_crs(rec100_pp)[1][[1]], proj)
 
 ################################################################################
 # Reconstructed to c(0,100) with listout
-rec <- reconstruct(admin, age=c(0,100), listout=TRUE,  model=model, plateperiod=TRUE)
+rec <- reconstruct(admin, age=c(0,100), listout=TRUE,  model=model, validtime=TRUE)
 
 
 # high-level attributes
@@ -68,7 +68,7 @@ expect_equal(rec[[2]], rec100)
 ################################################################################
 # Reconstructed to c(0,100) with listout=FALSE
 expect_error(
-	rec <- reconstruct(admin, age=c(0,100), listout=FALSE,  model=model, plateperiod=TRUE)
+	rec <- reconstruct(admin, age=c(0,100), listout=FALSE,  model=model, validtime=TRUE)
 )
 
 
@@ -94,7 +94,7 @@ expect_equal(st_crs(rec0)[1][[1]], proj)
 
 ## ################################################################################
 # Reconstructed to 100 
-rec100 <- reconstruct(admin, age=100, model=model, plateperiod=TRUE)
+rec100 <- reconstruct(admin, age=100, model=model, validtime=TRUE)
 
 
 
@@ -108,8 +108,8 @@ expect_equal(colnames(rec0), colnames(rec100))
 expect_equal(st_crs(rec100)[1][[1]], proj)
 
 ################################################################################
-# Reconstructed to 100 with plateperiod=FALSE
-rec100_pp <- reconstruct(admin, age=100, model=model, plateperiod=FALSE)
+# Reconstructed to 100 with validtime=FALSE
+rec100_pp <- reconstruct(admin, age=100, model=model, validtime=FALSE)
 
 
 expect_inherits(rec100_pp, "Spatial")
@@ -124,7 +124,7 @@ expect_equal(st_crs(rec100_pp)[1][[1]], proj)
 
 ################################################################################
 # Reconstructed to c(0,100) with listout
-rec <- reconstruct(admin, age=c(0,100), listout=TRUE,  model=model, plateperiod=TRUE)
+rec <- reconstruct(admin, age=c(0,100), listout=TRUE,  model=model, validtime=TRUE)
 
 
 # high-level attributes
@@ -138,5 +138,5 @@ expect_equal(rec[[2]], rec100)
 ################################################################################
 # Reconstructed to c(0,100) with listout=FALSE
 expect_error(
-	rec <- reconstruct(admin, age=c(0,100), listout=FALSE,  model=model, plateperiod=TRUE)
+	rec <- reconstruct(admin, age=c(0,100), listout=FALSE,  model=model, validtime=TRUE)
 )