+++ tests/offline/test_offline_sp_points.R
@@ -47,11 +47,11 @@ expect_equal(st_crs(rec0)[1][[1]], "WGS 84")
 
 ## ################################################################################
 # Reconstructed to 100 
-rec100 <- reconstruct(spp, age=100, model=model, plateperiod=TRUE)
+rec100 <- reconstruct(spp, age=100, model=model, validtime=TRUE)
 
 
 # the data.frame method for cross checking
-recDF<- reconstruct(dmat[, c("lng", "lat")], age=100, model=model, plateperiod=TRUE)
+recDF<- reconstruct(dmat[, c("lng", "lat")], age=100, model=model, validtime=TRUE)
 
 
 expect_inherits(rec100, "Spatial")
@@ -76,8 +76,8 @@ expect_equal(part, full[rownames(part),])
 expect_equal(st_crs(rec100)[1][[1]], "WGS 84")
 
 ################################################################################
-# Reconstructed to 100 with plateperiod=FALSE
-rec100_pp <- reconstruct(spp, age=100, model=model, plateperiod=FALSE)
+# Reconstructed to 100 with validtime=FALSE
+rec100_pp <- reconstruct(spp, age=100, model=model, validtime=FALSE)
 
 
 expect_inherits(rec100_pp, "Spatial")
@@ -97,7 +97,7 @@ expect_equal(st_crs(rec100_pp)[1][[1]], "WGS 84")
 
 ################################################################################
 # Reconstructed to c(0,100) with listout
-rec <- reconstruct(spp, age=c(0,100), listout=TRUE,  model=model, plateperiod=TRUE)
+rec <- reconstruct(spp, age=c(0,100), listout=TRUE,  model=model, validtime=TRUE)
 
 
 # high-level attributes
@@ -112,7 +112,7 @@ expect_equal(rec[[2]], rec100)
 ################################################################################
 # Reconstructed to c(0,100) with listout=FALSE
 expect_error(
-	rec <- reconstruct(spp, age=c(0,100), listout=FALSE,  model=model, plateperiod=TRUE)
+	rec <- reconstruct(spp, age=c(0,100), listout=FALSE,  model=model, validtime=TRUE)
 )
 
 
@@ -143,11 +143,11 @@ expect_equal(st_crs(rec0)[1][[1]], proj)
 
 ## ################################################################################
 # Reconstructed to 100 
-rec100 <- reconstruct(spp, age=100, model=model, plateperiod=TRUE)
+rec100 <- reconstruct(spp, age=100, model=model, validtime=TRUE)
 
 
 # the data.frame method for cross checking
-recDF<- reconstruct(dmat[, c("lng", "lat")], age=100, model=model, plateperiod=TRUE)
+recDF<- reconstruct(dmat[, c("lng", "lat")], age=100, model=model, validtime=TRUE)
 
 
 expect_inherits(rec100, "Spatial")
@@ -172,8 +172,8 @@ expect_equal(part, full[rownames(part),])
 expect_equal(st_crs(rec100)[1][[1]], proj)
 
 ################################################################################
-# Reconstructed to 100 with plateperiod=FALSE
-rec100_pp <- reconstruct(spp, age=100, model=model, plateperiod=FALSE)
+# Reconstructed to 100 with validtime=FALSE
+rec100_pp <- reconstruct(spp, age=100, model=model, validtime=FALSE)
 
 
 expect_inherits(rec100_pp, "Spatial")
@@ -193,7 +193,7 @@ expect_equal(st_crs(rec100_pp)[1][[1]], proj)
 
 ################################################################################
 # Reconstructed to c(0,100) with listout
-rec <- reconstruct(spp, age=c(0,100), listout=TRUE,  model=model, plateperiod=TRUE)
+rec <- reconstruct(spp, age=c(0,100), listout=TRUE,  model=model, validtime=TRUE)
 
 
 # high-level attributes
@@ -208,7 +208,7 @@ expect_equal(rec[[2]], rec100)
 ################################################################################
 # Reconstructed to c(0,100) with listout=FALSE
 expect_error(
-	rec <- reconstruct(spp, age=c(0,100), listout=FALSE,  model=model, plateperiod=TRUE)
+	rec <- reconstruct(spp, age=c(0,100), listout=FALSE,  model=model, validtime=TRUE)
 )
 
 
@@ -240,11 +240,11 @@ expect_equal(st_crs(rec0)[1][[1]], proj)
 
 ## ################################################################################
 # Reconstructed to 100 
-rec100 <- reconstruct(spp, age=100, model=model, plateperiod=TRUE)
+rec100 <- reconstruct(spp, age=100, model=model, validtime=TRUE)
 
 
 # the data.frame method for cross checking
-recDF<- reconstruct(dmat[, c("lng", "lat")], age=100, model=model, plateperiod=TRUE)
+recDF<- reconstruct(dmat[, c("lng", "lat")], age=100, model=model, validtime=TRUE)
 
 
 expect_inherits(rec100, "Spatial")
@@ -269,8 +269,8 @@ expect_equal(part, full[rownames(part),])
 expect_equal(st_crs(rec100)[1][[1]], proj)
 
 ################################################################################
-# Reconstructed to 100 with plateperiod=FALSE
-rec100_pp <- reconstruct(spp, age=100, model=model, plateperiod=FALSE)
+# Reconstructed to 100 with validtime=FALSE
+rec100_pp <- reconstruct(spp, age=100, model=model, validtime=FALSE)
 
 
 expect_inherits(rec100_pp, "Spatial")
@@ -290,7 +290,7 @@ expect_equal(st_crs(rec100_pp)[1][[1]], proj)
 
 ################################################################################
 # Reconstructed to c(0,100) with listout
-rec <- reconstruct(spp, age=c(0,100), listout=TRUE,  model=model, plateperiod=TRUE)
+rec <- reconstruct(spp, age=c(0,100), listout=TRUE,  model=model, validtime=TRUE)
 
 
 # high-level attributes
@@ -305,5 +305,5 @@ expect_equal(rec[[2]], rec100)
 ################################################################################
 # Reconstructed to c(0,100) with listout=FALSE
 expect_error(
-	rec <- reconstruct(spp, age=c(0,100), listout=FALSE,  model=model, plateperiod=TRUE)
+	rec <- reconstruct(spp, age=c(0,100), listout=FALSE,  model=model, validtime=TRUE)
 )